requirements.txt
aiohttp~=3.7.4.post0
cheroot~=8.5.2
ecdsa~=0.17
Flask~=2.0.2
Jinja2~=3.0.3
jsonschema~=4.2.1
parsedatetime~=2.6
python-dateutil~=2.8.2
pytz~=2021.3
PyYAML~=6.0
requests~=2.26.0
rfc3339-validator~=0.1.4